CDR-1615 Refactor allow-template overwrite (#1440)

+++ service/src/main/java/org/ehrbase/service/TemplateDBStorageService.java
@@ -28,6 +28,9 @@ import org.ehrbase.repository.CompositionRepository;
 import org.ehrbase.repository.TemplateStoreRepository;
 import org.ehrbase.util.TemplateUtils;
 import org.openehr.schemas.v1.OPERATIONALTEMPLATE;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -36,15 +39,35 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class TemplateDBStorageService implements TemplateStorage {
 
+    private static final String PROP_ALLOW_TEMPLATE_OVERWRITE = "ehrbase.template.allow-overwrite";
+
+    private static final Logger log = LoggerFactory.getLogger(TemplateDBStorageService.class);
+
     private final CompositionRepository compositionRepository;
 
     private final TemplateStoreRepository templateStoreRepository;
 
+    private final boolean allowTemplateOverwrite;
+
     public TemplateDBStorageService(
-            @Lazy CompositionRepository compositionRepository, TemplateStoreRepository templateStoreRepository) {
+            @Lazy CompositionRepository compositionRepository,
+            TemplateStoreRepository templateStoreRepository,
+            @Value("${" + PROP_ALLOW_TEMPLATE_OVERWRITE + ":false}") boolean allowTemplateOverwrite) {
 
         this.compositionRepository = compositionRepository;
         this.templateStoreRepository = templateStoreRepository;
+        this.allowTemplateOverwrite = allowTemplateOverwrite;
+
+        if (allowTemplateOverwrite) {
+            log.warn(
+                    "Template overwriting is enabled, this is not recommended for production use and can lead to unexpected behavior, consider disabling {}",
+                    PROP_ALLOW_TEMPLATE_OVERWRITE);
+        }
+    }
+
+    @Override
+    public boolean allowTemplateOverwrite() {
+        return allowTemplateOverwrite;
     }
 
     @Override
@@ -59,10 +82,24 @@ public class TemplateDBStorageService implements TemplateStorage {
 
     @Override
     public TemplateMetaData storeTemplate(OPERATIONALTEMPLATE template) {
-        if (findUuidByTemplateId(TemplateUtils.getTemplateId(template)).isEmpty()) {
+        String templateId = TemplateUtils.getTemplateId(template);
+        if (findUuidByTemplateId(templateId).isEmpty()) {
             return templateStoreRepository.store(template);
         } else {
-            checkUsage(template.getTemplateId().getValue(), "update");
+            boolean templateUsed = compositionRepository.isTemplateUsed(templateId);
+            if (templateUsed) {
+                if (allowTemplateOverwrite) {
+                    log.warn(
+                            "Updating template {} that is in use by at least one composition because {} is enabled",
+                            templateId,
+                            PROP_ALLOW_TEMPLATE_OVERWRITE);
+                } else {
+                    // There are compositions using this template -> Return list of uuids
+                    throw new UnprocessableEntityException(
+                            "Cannot update template %s since it is used by at least one composition"
+                                    .formatted(templateId));
+                }
+            }
             return templateStoreRepository.update(template);
         }
     }
@@ -70,15 +107,6 @@ public class TemplateDBStorageService implements TemplateStorage {
     @Override
     public Optional<TemplateMetaData> readTemplate(String templateId) {
         return templateStoreRepository.findByTemplateId(templateId);
-    }
-
-    private void checkUsage(String templateId, String operation) {
-
-        if (compositionRepository.isTemplateUsed(templateId)) {
-            // There are compositions using this template -> Return list of uuids
-            throw new UnprocessableEntityException("Cannot %s template %s since it is used by at least one composition"
-                    .formatted(operation, templateId));
-        }
     }
 
     private void checkUsages() {
@@ -99,9 +127,11 @@ public class TemplateDBStorageService implements TemplateStorage {
     @Override
     public void deleteTemplate(String templateId) {
 
-        // Check if template is used anymore
-        checkUsage(templateId, "delete");
-
+        if (compositionRepository.isTemplateUsed(templateId)) {
+            // There are compositions using this template -> Return list of uuids
+            throw new UnprocessableEntityException(
+                    "Cannot delete template %s since it is used by at least one composition".formatted(templateId));
+        }
         templateStoreRepository.delete(templateId);
     }
